requirements_macos.txt
# HeartMuLa Studio - macOS Requirements
# Optimized for PyInstaller bundling with Apple Metal GPU support

# Core Dependencies
fastapi>=0.115.2
uvicorn>=0.27.0
sqlmodel>=0.0.16
python-multipart>=0.0.18
python-dotenv>=1.0.0
requests>=2.32.2
aiofiles>=23.2.1
pydantic>=2.6.1
httpx>=0.27.0
sse-starlette>=2.0.0
huggingface_hub>=0.20.0

# PyTorch for Apple Metal
# Use conda or pip install torch for macOS with Metal support:
# pip install torch torchvision torchaudio
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# HeartLib - AI Music Generation Engine
heartlib @ git+https://github.com/HeartMuLa/heartlib.git

# Transformers
transformers>=4.35.0

# Audio Processing
soundfile>=0.12.0
librosa>=0.10.0

# Optional: Triton for torch.compile optimization (Linux/macOS Intel)
# Required for optimal performance with torch.compile
# Note: Not available on Apple Silicon - will be skipped gracefully
triton>=2.0.0; platform_machine != "arm64"

# Optional: Advanced ML dependencies
# These are optional for the application and imported with try/except in the codebase
# Uncomment only if you encounter import errors during runtime
# mamba-ssm  # Uncomment if heartlib requires mamba-ssm for specific features
# tensorboard  # Uncomment if you need TensorBoard logging capabilities

# PyInstaller for macOS App Bundle
pyinstaller>=6.0,<7.0

# PyWebView for native macOS UI
pywebview>=4.0
